#10072: Move test files to TTNN
