+++ tests/ttnn/unit_tests/operations/backward/test_backward_atanh.py
@@ -5,7 +5,7 @@
 import torch
 import pytest
 import ttnn
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import data_gen_with_range, compare_pcc
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_with_range, compare_pcc
 
 
 @pytest.mark.parametrize(
@@ -16,13 +16,13 @@ from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs i
         (torch.Size([1, 3, 320, 384])),
     ),
 )
-def test_bw_asinh(input_shapes, device):
-    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
+def test_bw_atanh(input_shapes, device):
+    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, required_grad=True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
 
-    pyt_y = torch.asinh(in_data)
+    pyt_y = torch.atanh(in_data)
 
-    tt_output_tensor_on_device = ttnn.asinh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.atanh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
